Move GameEngine into frontend
