+++ src/main/java/backend/Race.java
@@ -1,5 +1,7 @@
 package backend;
 
+import frontend.GameEngine;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
