+++ src/main/java/frontend/controllers/EditTeamController.java
@@ -1,6 +1,6 @@
 package frontend.controllers;
 
-import backend.GameEngine;
+import frontend.GameEngine;
 import backend.Season;
 import backend.Team;
 import javafx.event.ActionEvent;
