+++ src/main/java/frontend/controllers/HomeController.java
@@ -1,6 +1,6 @@
 package frontend.controllers;
 
-import backend.GameEngine;
+import frontend.GameEngine;
 import backend.Season;
 import backend.Team;
 import javafx.fxml.FXML;
