+++ src/main/java/frontend/controllers/PrepareRaceController.java
@@ -1,6 +1,6 @@
 package frontend.controllers;
 
-import backend.GameEngine;
+import frontend.GameEngine;
 import backend.Season;
 import backend.Setup;
 import backend.Strategy;
