+++ src/main/java/frontend/controllers/StandingsController.java
@@ -1,6 +1,6 @@
 package frontend.controllers;
 
-import backend.GameEngine;
+import frontend.GameEngine;
 import backend.Season;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
