+++ src/main/java/frontend/controllers/TransferOfferController.java
@@ -1,7 +1,7 @@
 package frontend.controllers;
 
 import backend.Driver;
-import backend.GameEngine;
+import frontend.GameEngine;
 import backend.Season;
 import backend.Staff;
 import backend.Team;
